There should never have been a header file; its private
No user should ever be aware of this class...
