+++ opencog/persist/guile/PersistSCM.cc
@@ -21,10 +21,48 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
+#ifndef _OPENCOG_PERSIST_SCM_H
+#define _OPENCOG_PERSIST_SCM_H
+
+#include <opencog/atomspace/AtomSpace.h>
+#include <opencog/atoms/base/Handle.h>
+#include <opencog/guile/SchemeModule.h>
+
+namespace opencog
+{
+/** \addtogroup grp_persist
+ *  @{
+ */
+
+class PersistSCM : public ModuleWrap
+{
+private:
+	void init(void);
+
+	Handle fetch_atom(Handle);
+	Handle fetch_incoming_set(Handle);
+	Handle fetch_incoming_by_type(Handle, Type);
+	Handle store_atom(Handle);
+	void load_type(Type);
+	void load_atomspace(void);
+	void store_atomspace(void);
+	void barrier(void);
+
+public:
+	PersistSCM(void);
+}; // class
+
+/** @}*/
+}  // namespace
+
+extern "C" {
+void opencog_persist_init(void);
+};
+
+#endif // _OPENCOG_PERSIST_SCM_H
+
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/guile/SchemePrimitive.h>
-
-#include "PersistSCM.h"
 
 using namespace opencog;
